refactor: [GNS-258] Improve repository structure
